Move testing strategies to their own file.
The `hypothesis` strategies for datasets and `Evaluator` objects are
bound to be used all over the place, so I've put them in
`tests/utils.py`.

The filename is deliberately generic in case there are other helper
functions we may need.
